// lib/types.ts
export type ProfileType = {
  id: number
  username: string
  email: string
  name: string
  slug: string
  nickname: string
  gender: string
  image: string
  dob: string
  website: string
  contact: string
  contact_email: string
  linkedin: string
  github: string
  address: string
  about: string
  is_portfolio_user: string
  resume_link: string
  is_active: string
  is_staff: string
  is_superuser: string
  date_joined: string
  last_login: string
  updated_at: string
}
// ...
export type BlogCategoryType = {
  id: number
  name: string
  slug: string
  created_at: string
  updated_at: string
}

export type TableOfContents = {
  id: string
  level: number
  heading: string
}

export type BlogType = {
  id: number
  slug: string
  title: string
  category?: BlogCategoryType
  image: string
  overview?: string
  content: string
  author: string
  tags?: string
  status: string
  order: number
  table_of_contents?: TableOfContents[]
  total_views: number
  total_likes: number
  user_liked: boolean
  created_at: string
  updated_at: string
}
// ...
export type ViewsType = {
  total_views: number
  total_likes: number
  liked: boolean
}

export type LikeStatusType = {
  liked: boolean
}